refactor stream operator<<; reset default numeric precision to 18

+++ include/anyrpc/xml/xmlwriter.h
@@ -42,7 +42,7 @@ public:
     //! Make the text more readable with indentation and linefeeds
     void SetPretty(bool pretty=true) { pretty_ = pretty; }
     //! Set the double format method. The default is no exponents for xmlrpc spec compatibility.
-    void SetScientificPrecision(unsigned precision=32)
+    void SetScientificPrecision(unsigned precision=18)
         { precision_ = std::min( 32u, precision); }
 
     //!@name Miscellaneous Member Functions
